Fix incorrect rootUrl by reversing resolveRoot search order
Previously if you opened a file in a directory like:

  /home/ryan/project/subproject/src/foo

It would search from the root down, and in my case, where I have a
clojure project with rootPattterns is ['project.clj'], and where
"project" is a monorepo whose root also has a `project.clj` it ends up
incorrectly picking `/home/ryan/project` instead of
`/home/ryan/project/subproject`.

imo, it makes more sense to search from the directory `dir` *up* (eg:
find the first parent directory of the file you are working on that
contains a file in `rootPatterns`

To be honest im not sure what the isParentFolder check was supposed
to do, but it seemed to just interfered, but i probably broke something.